major-projects/kmeans-image-compressor/requirements.txt
gunicorn
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
flask==3.1.2
pillow
scikit-image